LayoutTests/inspector/styles-source-offsets.html

<html>
<head>

<link rel="stylesheet" href="resources/styles-source-offsets.css" -->

<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="elements-tests.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        if (result === "refreshing")
            return;
        dumpArray(result);
        notifyDone();
    }
    evaluateInWebInspector("frontend_runTest", callback);
}


// Frontend functions.

function frontend_runTest(testController)
{
    testController.waitUntilDone();
    frontend_expandDOMSubtree(WebInspector.domAgent.document);

    testController.runAfterPendingDispatches(function() {
        function stylesCallback(testController, styles)
        {
            var matchedCSSRules = styles.matchedCSSRules;
            for (var i = 0; i < matchedCSSRules.length; ++i) {
                var rule = matchedCSSRules[i];
                frontend_dumpStyleData(testController, rule, false);
            }
            frontend_dumpStyleData(testController, styles.inlineStyle, true);
        }
        InspectorBackend.getStylesForNode(WebInspector.domAgent.document.body.id, stylesCallback.bind(null, testController));
    });
}

function frontend_dumpStyleData(testController, ruleOrStyle, shouldBail)
{
    var isRule = !!(ruleOrStyle.style);
    var style;
    var header = "";
    if (isRule) {
        if (ruleOrStyle.origin !== "") {
            if (shouldBail)
                testController.notifyDone(testController.results);
            return;
        }
        style = ruleOrStyle.style;
        header = ruleOrStyle.selectorText + ": [" + (ruleOrStyle.selectorRange ? ruleOrStyle.selectorRange.start + "-" + ruleOrStyle.selectorRange.end : "")  + "]";
    } else {
        style = ruleOrStyle;
        header = "element.style:";
    }
    testController.results.push(header + " [" + style.properties.startOffset  + "-" + style.properties.endOffset + "]");
    var allProperties = style.cssProperties;
    for (var i = 0; i < allProperties.length; ++i) {
        var property = allProperties[i];
        if (property.status === "style")
            continue;
        testController.results.push("['" + property.name + "':'" + property.value + "'" + (property.priority ? " !important" : "") + (property.parsedOk ? "" : " non-parsed") +"] @[" + property.startOffset + "-" + property.endOffset + "]");
    }
    if (shouldBail)
        testController.notifyDone(testController.results);
}

</script>

<style>

body.mainpage {
    text-decoration: none; /* at least one valid property is necessary for WebCore to match a rule */
    badproperty: 1badvalue1;
}

</style>
</head>

<body class="main1 main2 mainpage" onload="onload()" style="font-weight: normal; width: 80%">
<p>
Tests that proper data and start/end offset positions are reported for CSS style declarations and properties.
</p>

</body>
</html>
